fix(gateway): clarify destructive semantics
