+++ docs/dev/OBSERVABILITY.md
@@ -250,7 +250,8 @@ Destructive actions must log:
 
 Intent logs must make it clear which action is about to mutate state. Outcome logs must indicate success or failure.
 
-Gateway reconcile actions are destructive in this contract:
+Gateway reconcile actions currently marked `ActionSpec.destructive` must log
+their confirmation-gated intent and outcome:
 
 - `gateway.add`
 - `gateway.update`
